+++ app/src/androidTest/java/woowacourse/payments/component/PaymentCardTest.kt
@@ -32,6 +32,7 @@ class PaymentCardTest {
                 cardOwner = "뭉치",
                 password = "1234",
                 bankType = BankType.BC,
+                id = 1L,
             )
 
         // when
