+++ app/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.registercard.RegisterCardActivity"
+            android:name=".ui.cardform.CardFormActivity"
             android:exported="false"
             android:label="CardRegisterActivity"
             android:theme="@style/Theme.Androidpayments" />
