+++ app/src/main/java/woowacourse/payments/ui/cardform/component/CardNumberInputField.kt
@@ -1,19 +1,20 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.cardform.component
 
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
+import woowacourse.payments.ui.cardform.CardNumberVisualTransformation
 
 @Composable
-fun CardOwnerInputField(
+fun CardNumberInputField(
     text: String,
     onValueChange: (String) -> Unit,
     isError: Boolean,
@@ -22,30 +23,24 @@ fun CardOwnerInputField(
         modifier = Modifier.fillMaxWidth(),
         value = text,
         onValueChange = onValueChange,
-        label = { Text(stringResource(R.string.card_owner_label)) },
+        label = { Text(stringResource(R.string.card_number_label)) },
         placeholder = {
             Text(
-                text = stringResource(R.string.card_owner_placeholder),
+                text = stringResource(R.string.card_number_place_holder),
                 color = Color.Gray,
             )
         },
+        visualTransformation = CardNumberVisualTransformation(),
         singleLine = true,
-        supportingText = {
-            Row(modifier = Modifier.fillMaxWidth()) {
-                Spacer(Modifier.weight(1f))
-                Text(
-                    stringResource(R.string.card_owner_length, text.length),
-                )
-            }
-        },
+        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         isError = isError,
     )
 }
 
 @Preview(showBackground = true)
 @Composable
-private fun CardOwnerInputFieldPreview() {
-    CardOwnerInputField(
+private fun CardNumberInputFieldPreview() {
+    CardNumberInputField(
         text = "",
         onValueChange = { },
         isError = false,
