+++ app/src/main/java/woowacourse/payments/ui/cardform/component/CardOwnerInputField.kt
@@ -1,46 +1,51 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.cardform.component
 
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.KeyboardType
-import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 import woowacourse.payments.R
 
 @Composable
-fun PasswordInputField(
+fun CardOwnerInputField(
     text: String,
     onValueChange: (String) -> Unit,
     isError: Boolean,
 ) {
     OutlinedTextField(
-        modifier = Modifier.fillMaxWidth(0.5f),
+        modifier = Modifier.fillMaxWidth(),
         value = text,
         onValueChange = onValueChange,
-        label = { Text(stringResource(R.string.password_label)) },
+        label = { Text(stringResource(R.string.card_owner_label)) },
         placeholder = {
             Text(
-                text = stringResource(R.string.password_place_holder),
+                text = stringResource(R.string.card_owner_placeholder),
                 color = Color.Gray,
             )
         },
-        visualTransformation = PasswordVisualTransformation(),
         singleLine = true,
-        keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
+        supportingText = {
+            Row(modifier = Modifier.fillMaxWidth()) {
+                Spacer(Modifier.weight(1f))
+                Text(
+                    stringResource(R.string.card_owner_length, text.length),
+                )
+            }
+        },
         isError = isError,
     )
 }
 
 @Preview(showBackground = true)
 @Composable
-private fun PasswordInputFieldPreview() {
-    PasswordInputField(
+private fun CardOwnerInputFieldPreview() {
+    CardOwnerInputField(
         text = "",
         onValueChange = { },
         isError = false,
