+++ app/src/test/java/woowacourse/payments/registercard/CardNumberVisualTransformationTest.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.payments.ui.registercard.CardNumberVisualTransformation
+import woowacourse.payments.ui.cardform.CardNumberVisualTransformation
 
 class CardNumberVisualTransformationTest {
     private lateinit var visualTransformation: VisualTransformation
