+++ app/src/test/java/woowacourse/payments/registercard/ExpiryDateVisualTransformationTest.kt
@@ -5,7 +5,7 @@ import androidx.compose.ui.text.input.VisualTransformation
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
-import woowacourse.payments.ui.registercard.ExpiryDateVisualTransformation
+import woowacourse.payments.ui.cardform.ExpiryDateVisualTransformation
 
 class ExpiryDateVisualTransformationTest {
     private lateinit var visualTransformation: VisualTransformation
